+++ docs/REQUIREMENTS.md
@@ -78,7 +78,7 @@
 
 2차 피드백 이후 리팩터링 및 기능추가
 
-- [ ] refactor: 컨트롤러 분리
+- [x] refactor: 컨트롤러 분리
 - [ ] test: 자동차 이름, 시도횟수에 대한 테스트 추가
 
 ## 파일 구조
